Update CpuLoad.php add sys_getloadavg existence check
The [sys_getloadavg](https://www.php.net/manual/en/function.sys-getloadavg.php) is not implemented on Windows platforms. When it is called, it throws `Fatal error` (Call to undefined function sys_getloadavg()). There are some workarounds on the Internet but it seems easier just to check if it exists and provide no information on these platforms.